Work around issues with weak linkage and mmap/mprotect/munmap.
